Major update to compile expression file to Java.
This change establishes an entire pattern for using Bazel as the build system for this compiler to generate a file containing the intermediate Java source that the cup parser now generates instead of interpreting the calculation on the fly during parsing. Bazel rules are now configured to generate java jars from source *.calculator files which can now be executed hermetically using `$ bazel run compiled_calculator_binary`.