Implement a real REPL Terminal
Now the REPL backend depends on the Lanterna Terminal library https://github.com/mabe02/lanterna in order to support arrow keys for navigating through history of commands given in the REPL and for editing the current command.

There's still work needed on this REPL terminal implementation to support the terminal scrolling past the last line, and for pushing the cursor repositioning logic into the REPL implementation of the PrintStmt Node so that iteration works in the future.