Implement List initialization.
Lists are now supported in Claro. These are implemented as mutable Java ArrayLists under the hood. There's no support yet for accessing elements within these lists yet however, only initialization.

Currently assignment of lists to variables in the JAVA_SOURCE backend is a major hack. Types of variables are now determined by checking the instanceof the Expr node being assigned to the var. This is super flimsy because it's centralized in the AssignmentStmt.java class right now. The very next, most important TODO is going to be requiring all Expr impls to implement a method to get its type.