Implement list-subscripting.
There's now a generalized subscripting support in Claro for any nesting level. However, it's still just as hacky as ever in terms of the typing support. For now there's only support for Integer types in the JavaSource version and we still need to actually implement some sort of Expr::getType method everywhere so that we can decentralize type registration.