Implement len(...) to find the length of a List.
This included a required massive change to all of the arithmetic and comparison operators in order to support all combinations of Integer/Double in those operators. This was exremely painful and an ugly implementation for the INTERPRETED backend because all types needed to have instanceof called a million times. Perhaps in the future we'll move the compiler to Kotlin and take advantage of multiple dispatch support to help with some of this grossness.

Still need to implement proper type-checking support in a decentralized way throughout the AST.