(In-progress) Add type validation support!!
This will be the biggest change to Claro to date and is gonna be the leaping off point for all kinds of cool things down the road! Now Claro will be able to validate types of all expressions in all statements *before* even trying to execute them, warning of any found type mismatches before wasting your time and giving you nasty runtime error messages.

Claro will now support type inference on variables initialized with an expr on the rhs of the equality operator. Claro will also now require type declarations on variable declaration (also new itself) in order to make sure that the type of a variable is known on the first line that it appears in the program rather than having to potentially chase an undecided variable through arbitrary spaghetti code to infer its type. Claro takes the stance that it's not at all difficult for a programmer to type out the type of a declared variable that hasn't been initialized yet.

For posterity: the primary reason that explicit variable declaration with type annotations came into the language now was in order to logically support the initialization of a variable to an empty list of Foo. Previously this would have looked like `l = [];` but this is completely ambiguous considering at this point the next line (or the one hundreds of lines later that lets us understand its type) could just as equally be `l.append(1)` or `l.append(1.0)` or `l.append("one")`. Instead, now Claro requires the variable to be declared as `l: [int] = [];` to clarify that (of the previous options) only `l.append(1)` is valid. Claro could've marked this variable's type as undecided and added an intermediate crawl of the AST to find the code that makes this variable's actual type apparent, but that'd be such a ridiculous compiler performance cost for the gain of not typing out a few characters that I said fuck it, we're requiring type declaration in these situations.