Separate declaration and initialization into distinct concepts in Claro.
Now, to enable type checking of all kinds of situations including the finicky empty list assignment: `l = [];`, we'll now have var declarations separate from assignments. For now there's only support for a hardcoded var type which is assumed to be an int.

Also made a massive change to now validate that a variable has been initialized before usage, even inspecting branches to ensure that every branch prior to the identifier usage initializes the identifier to some value. I'm super hyped about this. For now it only supports if-statement chains, but doesn't support while loops because to support that I'd need to implement some sort of branch prediction to identify the cases where at compile time it can be determined that the while loop will enter the body at least once....that sounds not worth it for now.

Still need to come back to add support for parsing type annotations in the Parser.