Separate declaration and initialization into distinct concepts in Claro.
Now, support type checking of all kinds of situations including the finicky empty list assignment: `l = [];`, we now have var declarations separate from assignments.

Also added the additional `type` Stmt to Claro which for now is a hacked, lazy version of what it will end up being long term. For now, it simply prints a string repr of the type of the given expr which is known at compile time. This is not as strong as what we'll want long term which is to have the ability to have the meta type Type included in the runtime of the language itself as a builtin type in order to allow for high order programming. This is skipped for now since I'd want to simply have the JavaSource impl depend on the same Types.java file that the compiler impl depends on, but I genuinely cannot figure out how to have our genrule use javac on multiple java files yet....