Implement Function Calls!!!! :D:D
Now you can call functions...in the JAVA_SOURCE backend as well!!

Note that in this commit I also found and fixed a bug with handling scoping where the final global scope was never being exited after either the type-checking phase or the execution phase. The main side effect of this was that the "unused" checking was no longer working in the global scope, though it was working in nested scopes.