Support for higher order functions!
Claro now allows assignment of functions to variables which allows for passing those functions to other functions where they can be called later, this is a higher order function! This change also added syntax for defining the type of functions as follows:

Functions:
  - function<singleArgType -> returnType>
  - function<|argType1, argType2, ...| -> returnType>
Consumer Functions:
  - consumer<singleArgType>
  - consumer<argType1, argType2, ...>
Provider Functions:
  - provider<returnType>

A major enabling change here is the fact that all user provided names have been removed from the typing system as of now. So, function names and function arg names are no longer relevant to the type of a function of any sort. These names are essentially considered user level comments for now. Likely in the future these names related to function parts will become useful for things like named params allowing the passing of args in any order, however, that is still not related to the actual *type* of the function, it's just external metadata about the function's type, however that information is not part of the function's type itself. So we'll get to it when we need it. (We'll also need the function's name for when we want to implement recursion since the JAVA_SOURCE backend will need to know the original class name that it should be deferring to even when the function reference may be under a totally different variable name...).