Support for Tuples!!
This is a MASSIVE change that includes the addition of Tuples and some updates to initialization of Structs.

The main changes here are all related to supporting the undecidable type problem related to subscripting a Tuple since each index of a Tuple may contain values of different types, but yet the subscript Expr value is not known until runtime. As a result, Claro now includes the Cast Expr where the programmer can assert an expected type to inject more information that will be trusted at compile-time but validated at runtime.