Struct Definitions w/ Auto-generated Builders!
This CL is mostly a hack, but it implements support for instantiating user-defined Struct instances using an auto-generated Builder.

The hack in this CL needs to be immediately fixed in followups to this. The main issue is that all of the setter calls on the auto-generated builder and the build() call itself are completely hardcoded into the grammar itself which is literal insanity. We need actual type support for object instances that have associated methods. This is why this has been punted, because Claro currently has no support for objects or methods at all. So if I'm able to actually implement this for builders it's actually a perfect segway for plowing directly ahead into user-defined-classes.