Implement multi-phase compiler preprocessing phases!
Claro now has the following phases:
  1. "Type-Discovery"  <- **new in this CL**
  2. "Type-Validation"
  3. "Unused/Uninitialized-Checking"
  4. "Code-Generation"

From now on all types detected at the Parser are actually wrapped in a TypeProvider where the Type will later be resolved.

Some work is still going to be needed to allow referencing types defined in arbitrary order throughout the file, however, for now as long as the references come strictly after the definition, struct type references are working as expected.