Implement first scaffolding for REPL-site.
Little mental break from the hardcore compiler stuff. This is a sub-project I'll work on for fun/for the sake of better Claro evangelizing. This is the initial groundwork for a server to run a REPL website for Claro. Eddie will build the frontend and I'll build out a backend. This should be interesting because I'll want to support multiple concurrent users and I'll want to implement sessions to maintain state, and determine some scheme for extending/ending sessions as they're not used anymore.

Anyways, for now, this particular CL simply takes some params from the url and throws it into the parser. The most significant change is updating the PrintStmt.java Node so that we can inject a delegate function to handle printing. This delegate function in this REPL-site just collects strings to be printed so that they can be returned over the network to the browser.