Function args now *hide* same-named variables in outer scope.
This change basically ensures that the following situation is possible:

```
var testVarHiding: string = "this is a string";
consumer foo(testVarHiding: int) {
  print(testVarHiding++);
}
print(testVarHiding); # Still prints "this is a string".
```

Also added a debug function `$debugscope();` intended for use in the REPL where we can dump out the ScopedHeap to investigate what's going on. it's not perfect and I'll probably remove it any day since it's not formally part of the language.