Fix Recursion to work on all procedure types.
This ended up being a surprisingly tiny change, however, it was very tricky to figure out how to get it working. The final realization was that recursion required a reference to self by hiding the function reference variable in the outer scope of the same name with a reference to `this` since that variable in the outer scope wasn't yet existing (according to generated Java's scoping rules).