Enable Providers and Functions to Have Multiple Return Statements!
I'm super proud of this change. This one was actually most challenging to implement for the Interpreted backend. Thinking about how to implement returns w/o access to a GOTO statement was a trip. More detailed documentation of how I did that is given in the ReturnStmt class.

TODO(steving) I need to follow this up to come up with a clean way to disallow return statements outside of procedure bodies.