Disallow `return` Outside Procedures, and Require a `return` Along Every Procedure Branch.
Now Claro actually fully supports Procedure returns! This was a super clever hack to leverage the existing branch analysis logic for variable initialization checking to check for ReturnStmts along all possible codepaths through a procedure. This change creates a hidden uninitialized variable "$RETURNS" as the first step inside the procedure body's scope, and marks that hidden variable initialized every time that a ReturnStmt is found, then as the final step before exciting the procedure body's scope, we check that the hidden variable is initialized. This allows us to guarantee that every path through this procedure definitely has some return. We also unfortunately add a check before every Stmt is type checked to validate that the hidden variable is *not* initialized in the immediate scope level otherwise we know that this is definitely unreachable code (this last bit may not be the most efficient thing in the world but I don't have a much better approach than this).