Initial Commit for LambdaExpr!!!
This is probably the proudest thing that I've built in Claro yet. In a single long day of work I was able to get LambdaExprs implemented! This builds very heavily off of the logic that was implemented for ProcedureDefinitionStmts and uses some clever hacks to make sure that we still maintain all our nice features such as return branch inspection. Lambdas are very similar to other procedures except that they grab a closure of the variables defined before them in their scope. I wish I could say that this was similarly difficult compared to other procedure types, but I think that this was by far more complicated.

I still have more work to do in the grammar. Currently only lambda functions of a single arg exist, no other lambda types. Fast follow on that but I'm submitting now before I lose all this work.