Implement Format Strings Over Arbitrary Expressions!
Claro now supports format strings to allow essentially parameterizing a string based on arbitrary string type format expressions. This is literally just syntactic sugar to make it dumb easy to format strings in this language. There's some weird IntelliJ syntax highlighting going on since I guess this isn't necessarily a feature they expect in a language...but hey that's dope that I've gone beyond what IntelliJ even expects to need to support for custom languages!

Also added a small fix to allow special chars to be printed as expected in both the REPL and the JAVA_SOURCE backend.