Implement `exec(...)` in StdLib To Interpret Arbitrary Code in String.
This is obviously a generally dangerous thing, but it also leads to some really awesome fun with a language. Considering every other form of security in a programming language really just involves expecting programmers to do smart things, I'm not that sad about supporting exec from "security" concerns...just don't use it in prod code and you'll be happy lol.

The awesome thing about this is that it demonstrates the power of implementing the compiler backends themselves in the same language as the generated target source. This allows some super cool dogfooding, and the fact that this was a one night feature to implement proves the value of developing a compiler in this manner before transitioning to the more annoying thing of emitting bytecode.