Rich Error Messages!
This was easily the most obnoxious feature that I've ever added to Claro, but it's turned out to be an incredibly powerful one. Now Claro supports rich error messages to find as many syntax and type errors as possible all in one go and to notify the user all at once. This is arguably below table stakes for any programming language.

Compared to Java, for example, Claro pushes the needle forward in this regard in its tracking of full, potentially multi-line, spans to indicate very precisely the exact expression that caused a type mismatch as opposed to simply pointing to the nearest operator or parenthesis.

For example, for the Java program with type mismatch:
```
private void foo(String s) {...}
public static void main(String...args) {
    // btw I'm very intentionally using some pathalogical formatting, to show how Claro's error messaging is unique
  foo(10
     *
         9.5);
 }
```
You get this error message:
```
src/Error.java:6: error: incompatible types: double cannot be converted to String
     *
     ^
```
Which isn't very useful since it seems very arbitrary that the problem is related to the * operator and not the double expression...

So in Claro for the analogous program I get this message:
```
error.claro:8: Invalid type: found <float>, but expected <string>.
foo(10
  *
       9.5);
  ^^^^^^^^
```