Tweaks to Support Automated Releases To Riju
A few fixes to ensure that the build artifacts automatically created for Riju release can successfully work without change to the existing Riju onboarding configuration, namely ensuring that the Conditions.claro file is always available and builds to an intermediate Conditions.java before execution. This is an ugly hack but it takes so long to get things merged to Riju that I'm going to just leave this as is and deal with it.

The other small bit is the addition of a CLARO_VERSION.txt. The first line of this file is used simply as the version identifier that's printed in the REPL. This will be auto-updated by the GitHub Action to whatever the current release will be named so that I can easily go to Riju and determine which build is deployed since it will be an unknown when changes will reach Riju.