Support Module Composition!
This is solidifying into the heart and soul of Claro as a language. Claro now supports rich Module Composition semantics so that modules can declare their dependency on other modules. This comes with recursive validation that keys are not rebound and also comes with handy short-circuiting logic built in by default such that modules that are re-used are a no-op. This simply helps prevent shooting yourself in the foot and having to fight a nasty fight to factor shared deps out into higher level modules for no reason. The idea here is that every logical unit can declare its deps right next to the usage without thinking that hard about it.

TODO:
  1. Compile-time Binding Validation on procedure calls within using blocks
  2. Optimize the transitive key bindings set aggregation logic to avoid N^2 logic. Simply following the same logic used for transitive module dep aggregation using cached short-circuiting will return it to O(N).