Support lazy graph node provider injection!
Claro does this by allowing nodes to @ reference nodes but referencing them in a scope that statically asserts expectation of a provider<Foo> instead of just Foo. This conceptually exists to allow a graph node to decide whether or not to do something like write to a database based on the result of another async operation.

As far as I can tell Claro now has full parity with Google's Java Producers Framework (which is analogous to the external Dagger Producers). But Claro actually improves on both cases since Claro statically forbids deadlocking!