Implement Graph Consumers!
Claro now has support for Graph Consumers to round out full basic support for Graph Procedures! This is the culmination of a massive lift, I'd increment the version numbering scheme from 0.1.x to 0.2.x to represent this...but I'm too lazy to fiddle with that.

One brief note to keep in mind right away that I'm sure I'll have to explain repeatedly in more depth in the future. I made a conscious tradeoff to make Graph Consumers follow a fire-and-forget model where no handle (future) to their work is returned because I do not want to abuse the notion of a consumer that's been established in Claro to this point. Consumers do not return values. Futures are representations of work yes, but more pragmatically they are representations of values that will be ready in the future. If Claro's Graph Consumers were to return a handle to their work, then they would need to be returning something like future<nothing> which is not a clear win by any stretch of the imagination. For now, I'll assume that if you want some sort of cancellable Future to be returned from your procedure, then you should construct your own nothing type (or perhaps Claro will provide a singleton nothing and forbid anyone else from constructing one) and explicitly return future<nothing> from a Graph Function instead. There's no perfectly clean solution here and I believe I've chosen the cleanest of the available options.