Implement Pipe Chain Stmt For Cleaner Function Composition
Claro's pipe chain statement models a data source flowing through a pipeline of transformations and then finally landing in a sink statement. The sink must be a terminating assignment to an existing variable, or a call to a consumer function.

Pipe chains can be thought of as some form of syntax sugar, however, they also carry with them the added benefit that should not be overlooked that they allow you the opportunity to forego thinking of a name for an intermediate result when you would otherwise have to just to be able to use an intermediate value in more than one place in the following expression. Given the fact that naming is considered one of the hardest problems in computer science, I think this is a worthwhile feature on that merit alone. The fact that it also allows avoiding the slow march to the right that typically arises from the typical use of composed procedure calls is also a big win. Since Claro aims to be built entirely around functions and not classes, I want to make using Claro functions very comfy.