Annoying tiny bug fix
I was a bit slapdash in modelling the reference of the generated intermediate variables for pipe stage's whose value is backreferenced more than once. Previously I just passed manual Java Source codegen of the name of the generated var but it turned out that with the scoping rules of unused var checking, leaving the declared variable with its usage unmarked in the scopedheap broke things, and made Claro think the variable was unused even though it's generated and definitely actually used. I replaced the manual codegen with deference to the IdentifierReference node literally just to get it to mark the variable used... this was like a 3 line fix but now I'll have to live with not having this fix represented in Riju for a couple months since I jumped the gun asking him to push the update... *facepalm*