First Step Towards Implementing Contracts!
It will probably be hard to see at this current stage, but Contracts are Claro's foundation for polymorphism in the language. Generic functions will be built around deferring to Contract implementations for the generic types applied to the function at the call site, and monomorphization will mean that these calls to the appropriate Contract Implementation will be a static dispatch to the correct function.

I'm extremely excited about this because this is the beginning of the language finally achieving the types of code reuse that can be achieved via OO's implementation or inheritance paradigm WITHOUT having to sully Claro with any of those terrible things! Data and functionality should be separate. Do not personify your data.