CLARO HAS GENERICS!!!! (Limited, still need a ton of cleanup)
Claro finally has its first look at Generics!!!! This is still quite limited and doesn't yet even perform all of the necessary type validations, but this needs to get submitted before my laptop catches fire because this is too important.

Claro performs static generic validation on the body of the generic function first, and then does type validation on every other stmt in the program and collects a list of all valid calls to the generic function and then during codegen "Monomorphizes" the generic function to have a static implementation for every unique set of args that it is called with. This is next level stuff! Lot's more still to come here.