Quality of Life Improvements
Added an append<T>([T], T) consumer function as a builtin (I still have an open TODO to figure out how best to implement the stdlib particularly in relation to features that require compiler intrinsic definitions that can't otherwise be written in the language, but that will come after multi-file modules are implemented).

Generic functions are no longer required to have a `requires` clause, which is useful in case the programmer doesn't want to call contracts in their function def.