Support Generics and Blocking-Generics On the Same Function
This change is yet another extension to Generics support so that we're one step closer to having full feature completion of generics over Functions. The final remaining piece is to enable to the use of Generic Functions as first class objects just like non-generic procedures have support for already. After I finish that, the template for extending these features all in one go for Consumers and Providers should all be in place.