Advent of Code Day 1 in Claro!!
Taking a break from hardcore stuff, I can't believe I'm able to solve a "real" coding challenge in Claro!

Had to add some new features to Claro to implement this problem, which feels insane. Added support for indexing into strings, and a new IsInputReadyExpr for access to a boolean `isInputReady()` provider func that returns true if there is a line of stdin available to read, and false otherwise.

I'll try to keep on with Advent of Code for as long as Claro keeps up! (Likely won't be very long as at the moment Claro is not near feature complete.)