Advent of Code Day 2 in Claro!!
Taking a break from hardcore stuff, I can't believe I'm able to solve a "real" coding challenge in Claro!

Had to add some new features to Claro to implement this problem as well. Finally added a long needed modulus operator `%`. I also noticed an issue with chaining an `==` operator after a binary math expr, since the math exprs were previously producing primitive results. Claro now takes the stance of converting everything back to an object so that I can just use `.equals()` on everything, instead of worrying about `==` vs `.equals`...this is probably something that will be considered a performance problem, so may change this in the future to allow math to be done with primitives only.

Wish list:
  - a builtin `match` statement
  - map/dict builtin type

I'll try to keep on with Advent of Code for as long as Claro keeps up! (Likely won't be very long as at the moment Claro is not near feature complete.)