Claro Now Supports Maps as a Builtin Type!
This is such a long time coming but I'm super happy this is finally here, and even happier that this was finally motivated by Advent of Code Day 3! It's really paying off to take the challenge of working through as much of Advent of Code as I can as it's finally giving me the motivation to get back from the advanced language features and back to laying more solid foundations for the basics.

Now claro supports type safe maps defined like the following:
`var m: {string: int} = {};`
or
`var m = {"foo": 1, "bar": 2};`

As is the philosophy in Claro, type annotations are only necessary where otherwise impossible to statically infer. I'll note here that I actually originally intended to use the notation `map<K, V>` for the map type literal, however, I thankfully realized last minute that this creates a new token in the language that would conflict with the well known functional `map` procedure. I'm using the current notation just to leave that name available.

You can subscript cleanly like so:
`print(m["foo"]);`
or
`m["bar"] = 100;`