First Step Towards Supporting a Builtin Set Type in Claro!
This is very early on in the Set support. You can only create a set with some initializer list of values, and you cannot remove or add any values later. There is also currently no support for any actual set operations such as union/intersection/etc. Coming soon when I get the chance.

For now submitting this as is to have at least some minimal functionality available for Advent of Code day 7!