Finalize Support for Recursive Type Alias Definitions!
The coolest thing that this change does is statically reject "Impossible Recursive Alias Types" that could not possibly be instantiated in a finite number of steps (and are therefore useless).

E.g.:
`alias ImpossibleType : tuple<int, ImpossibleType>`

Beyond that, this is a big change that deals with ensuring that recursive type aliases are properly supported in all cases. This addresses the subtly complex type equivalence that the following assignments should be allowed:

```
alias IntTreeNode: tuple<int, [IntTreeNode]>
var node1: IntTreeNode = (1, []);
var node2: IntTreeNode = (2, [node1]);
var node3: IntTreeNode = node2[1][0];  # This is the tricky assignment.
# var node3 = node2[1][0]; # This was already trivially supported since the rhs type is just accepted.
```
This was tricky because internally in the compiler, the type indexed into on the rhs is the `AliasSelfReferenceType` referenceing `IntTreeNode` which is different than `IntTreeNode` itself. So I needed to override the .equals() impl to do lazy recursive unwrapping. This also exposed some complexity in the JavaSource codegen where Java didn't "believe me" that the type recursion was progressing as known by Claro (since these recursive types have to be represented by constructs like `ClaroList<Object>` in generated code). So now, all CollectionSubscriptExpr's generate a cast to the *raw* type of the nested value w/ type generics erased because apparently Java will allow arbitrary nested generic types to be assigned to one another if you cast away the generics....I got very lucky that I can abuse these unsafe casts, because otherwise I'd have to abandon this approach (Recursive aliases would need to define a new class to recurse over in Java source - this would break the current codegen in many ways since currently the fact that the type was originally an alias is largely erased w/in Claro compilation).