Update create-new-riju-release.yml
Github Actions is using Bazelisk to keep Bazel up-to-date by default. Unfortunately I'm not doing this locally yet.. So, in order to make sure this is building with the same version of Bazel I'm using for local development, manually set the version to what I'm using locally here.

TODO(steving) In the future: upgrade to start using Bazelisk locally and then drop this.

Signed-off-by: Jason Steving <32336750+JasonSteving99@users.noreply.github.com>