Support Trashcan Assignment Stmt
This new statement allows you to dump a value "into the trash", indicating that you'll never make use of the value. This is primarily intended for use when calling some side effecting provider/function whose value you don't care about. But it can also be useful within procedure defs where you don't actually want to make use of a parameter value (yet).

This statement is intended to serve as an explicit indication that a value is being thrown away. This is "syntactic salt" (as opposed to "sugar") as this is intended to serve as a very visible indication that you should validate you're throwing this value away on purpose rather than on accident.

This has been long-awaited in the language. As Claro requires all variables to be used at least once after initialization, you previously needed to eventually have every single value in the program funnel down into a print statement directly or indirectly as `print(...)` was really the only statement in the entire language that could swallow a value (there was a hack where you could also fmt strings into a trash string and then both strings are technically "used" w/o printing, but obviously this was still disgustingly hacky).