Inline Type Signature in Lambda Exprs
Previously lambdas used in an unconstrained context required a cast to constrain the lambda's actual concrete type signature. This cast is undesirable from a cosmetic perspective because it begs whether the cast is happening at runtime (answer: it is not). In order to make this type annotation appear less like a runtime cast, Claro now supports inline type signatures on Lambda exprs:

Old:
```
var pred = (function<int -> boolean>) x -> x > 0;
```

New:
```
var pred = (x: int) -> boolean { return x > 0; };
```

Keep in mind, the intention here is not to try making a more concise syntax, but to make a syntax that more clearly illustrates what this annotation is actually being used for, and that it doesn't impose runtime overhead. This new syntax is particularly useful when passing a lambda expr to a function taking a generic first-class procedure arg where the type must be constrained at the call-site rather than in the generic callee's signature.