Generics Support for Consumer Functions!
This brings Consumers to parity with Functions so that you may fully make use of generics with each.

TODO: Still need to come back and add support for Consumers/Providers in Contract definitions before Providers can make use of generics by requiring a contract that would provider a value of the generic type for it to operate over and eventually return. This is extremely low priority as the usefulness of such a procedure is very limited.