Contract Functions Can Be Annotated Blocking & Utilize Generics!
This is a major step towards Contracts supporting the full set of procedural capabilities provided by the rest of the Claro language. More to come to fully flesh out Contracts but this is a great step in that direction!

TODO: Still need to come back and add support for Consumers/Providers in Contract definitions before Providers can make use of generics by requiring a contract that would provider a value of the generic type for it to operate over and eventually return. This is extremely low priority as the usefulness of such a procedure is very limited.