Completed Support for All Forms of Function Defs in Contracts!
Now, you can define functions and generic functions, while also mixing in blocking annotations and blocking-generics as desired.

TODO: Still need to come back and add support for Consumers/Providers in Contract definitions before Providers can make use of generics by requiring a contract that would provider a value of the generic type for it to operate over and eventually return. This is extremely low priority as the usefulness of such a procedure is very limited.