Improved Contract Implementation Inference
Previously the implementation had overlooked structurally nested Contract Procedure arg types. So, the following was broken:

```
contract Foo<T> {
  function do(t: [T]) -> [T];
}
implement Foo<int> {
  function do(t: [int]) -> [int] { ... }
}
var res = Foo::do([10]); # (Incorrect) Compiler Error: Missing impl `Foo<[int]>`
```

The above situation (and all other nested arg types) is now supported. In addition Claro also now only does type inference over the minimal number of args (from left to right) to be able to infer all Contract type params to know which implementation it should be deferring to. Add to this the fact that it now automatically detects when a contextual type assertion is needed and if it is present it can avoid doing inference on even more args. Claro's type inference is finding itself.

TODO: Still need to come back and add support for Consumers/Providers in Contract definitions before Providers can make use of generics by requiring a contract that would provider a value of the generic type for it to operate over and eventually return. This is extremely low priority as the usefulness of such a procedure is very limited.