Contracts Support Basic Consumers!
You can finally include Consumers in Contracts, including implementation and calling. Big improvement! Next up I'll be adding support for all generic consumer forms.

TODO: Still need to come back and add support for Consumers/Providers in Contract definitions before Providers can make use of generics by requiring a contract that would provider a value of the generic type for it to operate over and eventually return. This is extremely low priority as the usefulness of such a procedure is very limited.