Contracts Now Support All Forms of Consumers!
TODO: Still need to come back and add support for Providers in Contract definitions before top-level Providers can make use of generics by requiring a contract that would provide a value of the generic type for it to operate over and eventually return. This is extremely low priority as the usefulness of such a procedure is very limited.