Support for Oneof Type "Narrowing" and "Widening"!
Claro's oneofs are already closing in to be on par with TypeScript's union types! Now, if you have done an equality check on a oneof, Claro now takes advantage of the knowledge that, if that equality check were to ever succeed, then it means that the concrete variant type of the oneof has actually been "narrowed". So the following is now supported:

```
var intOrString: oneof<int, string> = getRandomIntOrString();

if (intOrString == "one") {
  # Can now treat the intOrString as a concrete `string` type, b/c the type is "narrowed".
  print(intOrString[0]);
  print(intOrString[1]);
  print(intOrString[2]);
}
```

This is only for convenience though, so the second that you would decide to try reassigning to the variable with a oneof type, if you assign to a variant type other than the one the scope has been narrowed to, from that line onwards Claro treats the type as having been "widened" back to the original oneof type. (This might seem weird, but this exists so that you essentially have some way to "opt-out" of the type narrowing when necessary).

```
var intOrString = getRandomIntOrString();

if (intOrString == 1) {
  # You can do math with `intOrString` b/c it's type was narrowed to int.
  var oneHundred: int = intOrString + 99;

  # From here the type has now been "widened" back to the original oneof<int, string>.
  intOrString = "one hundred";

  # This would now be a compile error for `+` op expecting int or float, but getting oneof<int, string>.
  # print(intOrString + 99);
}
```

Oneofs will still need more work to become the foundation on which all dynamic dispatch is achieved in the language, so much more to come here!