Support for `instanceof` Operator to Check Oneof Variant Type
This is a fairly simple op with the single use of checking whether a oneof is currently set to a particular variant type or not. Type "narrowing" applies here as with the previous oneof concrete typed expr equality checking.

This CL also adds support for providing heterogenous initializer lists for lists declared to contain elements of a oneof type.

Oneofs will still need more work to become the foundation on which all dynamic dispatch is achieved in the language, so much more to come here!