Initial Support for Ad-Hoc Polymorphism via Dynamic Dispatch Over Oneofs!!
This is an absolutely massive step that has been in the planning for well over a year! Claro now has the first steps towards rich support of ad-hoc polymorphism!

Now, if you make a call to a Contract function (consumers not yet supported, providers never will be by definition) by passing it a oneof<T1, ..., TN> where there is no implementation for that specific oneof<> but there IS support for all of T1, ..., TN, then Claro will generate a dynamic dispatch call to determine the specific concrete impl needed based on the runtime types of the passed oneof(s). This even covers nested types with a oneof deep in the nested structure in the position of one of the type params.

This will be the heart and soul of Claro's polymorphism story and should have all the power of Julia's multiple dispatch w/o any of the drawbacks of an incomprehensible dispatch rule set. It should have strictly more expressive power than an OO language like Java without any of the inheritance of implementation comprehensibility issues that you get in an OO paradigm.

There is still much polishing needed including support for Contract Consumers as well as all generic contract functions/consumers. Even more importantly Claro still needs to handle accepting subsets of oneofs in procedure arg positions, and should infer the most narrow oneof variant set for the output type when the output type is determined by the input types. Lots of work here still to come.